Add the functionality of coverting temperature unit
So basically you input a temperature in K, and it gives the same amount of energy in other units. It is essentially
multiply with a Boltzmann constant kB. One of the output for thermal unit is still kT, in which one needs to specify at
which temperture is kT based. The input T actually make the default to be 1.